Move PrimitiveTypeDataNode to the correct namespace
